// orthnova-offline/src/types.ts
export type UserRole = 'admin' | 'doctor' | 'receptionist' | 'store_manager';

export interface UserRow {
  id: string; // UUID
  user_id: string;
  password: string;
  role: UserRole;
  created_at: string;
}